Restructure the repo to make it more clear  & Add Makefile
