+++ logalizer/table.go
