+++ logalizer/top_commands.go
@@ -2,26 +2,16 @@ package logalizer
 
 import "strings"
 
-// TotlaBandwidthCmd ... returns total bandwidth in KB , MB and GB
-func TotalBandwidthCmd(filename *string) {
-	F := ReadFile(*filename)
-	size := ParseLog("size", F)
-	totalsize := TotalBandwidth(size)
-	TabelBandwidth(totalsize)
-}
-
 // TopIpCmd .. returns top 10 IPs
 func TopIpCmd(filename *string) {
-	F := ReadFile(*filename)
-	ips := ParseLog("ip", F)
+	ips := ParseLog(*filename, "ip")
 	ip := TopOccurr(ips)
 	TablePrint("IP", "Count", ip)
 }
 
 // TopIP2LocCmd returns top 10 IPs with their locations
 func TopIP2LocCmd(filename *string) {
-	F := ReadFile(*filename)
-	ips := ParseLog("ip", F)
+	ips := ParseLog(*filename, "ip")
 	ip := TopOccurr(ips)
 	TablePrintIP2Loc("IP", "Count", "Location", ip)
 }
@@ -29,8 +19,7 @@ func TopIP2LocCmd(filename *string) {
 // TopMehodCmd ... returns top 10 methods
 func TopMethodCmd(filename *string) {
 	var list []string
-	F := ReadFile(*filename)
-	methods := ParseLog("method", F)
+	methods := ParseLog(*filename, "method")
 	for _, method := range methods {
 		list = append(list, method)
 	}
@@ -41,8 +30,7 @@ func TopMethodCmd(filename *string) {
 // TopRequestCmd ... returns top 10 requests
 func TopRequestCmd(filename *string) {
 	var list []string
-	F := ReadFile(*filename)
-	requests := ParseLog("request", F)
+	requests := ParseLog(*filename, "request")
 	for _, request := range requests {
 		list = append(list, request)
 	}
@@ -53,8 +41,7 @@ func TopRequestCmd(filename *string) {
 // TopReferrerCmd returns top 10 referrers
 func TopReferrerCmd(filename *string) {
 	var list []string
-	F := ReadFile(*filename)
-	referrers := ParseLog("referrer", F)
+	referrers := ParseLog(*filename, "referrer")
 	for _, referrer := range referrers {
 		referrer = strings.Trim(referrer, "\" ")
 		list = append(list, referrer)
@@ -66,8 +53,7 @@ func TopReferrerCmd(filename *string) {
 // TopStatusCmd ... returns top 10 status codes
 func TopStatusCmd(filename *string) {
 	var list []string
-	F := ReadFile(*filename)
-	response := ParseLog("response", F)
+	response := ParseLog(*filename, "response")
 	for _, code := range response {
 		list = append(list, code[:3])
 	}
@@ -79,8 +65,7 @@ func TopStatusCmd(filename *string) {
 // TopAgentCmd returns top 10 agents
 func TopAgentCmd(filename *string) {
 	var list []string
-	F := ReadFile(*filename)
-	agents := ParseLog("agent", F)
+	agents := ParseLog(*filename, "agent")
 	for _, agent := range agents {
 		agent = strings.Trim(agent, "\" ")
 		// table width problem if length of agent is too long
@@ -93,15 +78,4 @@ func TopAgentCmd(filename *string) {
 	}
 	result := TopOccurr(list)
 	TablePrint("Agent", "Count", result)
-}
-
-func ExecuteCustomRegex(regex *string, filename *string) {
-	var list []string
-	F := ReadFile(*filename)
-	matches := ParseCustom(*regex, F)
-	for _, match := range matches {
-		list = append(list, match)
-	}
-	result := TopOccurr(list)
-	TablePrint("Match", "Count", result)
 }
